Move file management to the RealmVersions file.
