+++ examples/ios/swift/Migration/RealmFile_v2.swift
@@ -30,7 +30,7 @@ struct RealmFile {
 
     static func create() {
         // Wipe old data
-        let v2Url = RealmNames.v2.url(clean: true)
+        let v2Url = RealmVersion.v2.destinationUrl(clean: true)
 
         // Create new data
         let person1 = Person(value: ["John Doe", 42])
